gateway/requirements.txt
fastapi
pydantic
pydantic_core
python-dotenv
uvicorn
pika
httpx
PyJWT
requests
python-multipart